requirements.txt
# Core packages
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0.1

# Deep learning
torch==2.0.1

# Web application
streamlit==1.25.0